+++ packages/inputs/src/types/control-value.ts
@@ -1,3 +1,3 @@
 import { ObjectWithPaths } from '@under-control/core';
 
-export type ControlValue = ObjectWithPaths | number | string | null;
+export type ControlValue = ObjectWithPaths | number | string | null | undefined;